requirements.txt
pyserial
numpy
pandas
scikit-learn
tensorflow
everywhereml